requirements.txt
fastapi
uvicorn[standard]
requests
beautifulsoup4
textblob
gspread
oauth2client
python-dotenv